Fix issues with int/bool/long.

+++ src/Simulation/Simulators/SparseSimulator2/Dump.cs
@@ -10,7 +10,7 @@ using Microsoft.Quantum.Simulation.Core;
 #nullable enable
 namespace Microsoft.Quantum.Simulation.Simulators
 {
-    using QubitIdType = System.IntPtr;
+    using QubitIdType = System.UInt32;
 
     public partial class SparseSimulator2
     {
